Rework the IO classes
