+++ optaplanner-benchmark/src/main/java/org/optaplanner/benchmark/impl/result/SolverBenchmarkResult.java
@@ -35,7 +35,7 @@ import org.optaplanner.core.api.score.Score;
 import org.optaplanner.core.api.solver.Solver;
 import org.optaplanner.core.config.solver.EnvironmentMode;
 import org.optaplanner.core.config.solver.SolverConfig;
-import org.optaplanner.core.impl.io.jaxb.JaxbIO;
+import org.optaplanner.core.impl.io.jaxb.GenericJaxbIO;
 import org.optaplanner.core.impl.score.definition.ScoreDefinition;
 
 /**
@@ -259,7 +259,7 @@ public class SolverBenchmarkResult {
     }
 
     public String getSolverConfigAsHtmlEscapedXml() {
-        JaxbIO<SolverConfig> xmlIO = new JaxbIO<>(SolverConfig.class);
+        GenericJaxbIO<SolverConfig> xmlIO = new GenericJaxbIO<>(SolverConfig.class);
         StringWriter stringWriter = new StringWriter();
         xmlIO.write(solverConfig, stringWriter);
         String xml = stringWriter.toString();
